requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
py-cord==2.7.1
distro==1.9.0
faiss-cpu==1.13.2
frozenlist==1.8.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.2
ollama==0.6.1
orjson==3.11.6
packaging==25.0
Pillow>=10.0.0
propcache==0.4.1
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pyte>=0.8.2
pypdf==6.6.2
python-dotenv==1.2.1
python-telegram-bot==22.6
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.1
sniffio==1.3.1
SQLAlchemy==2.0.46
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
